(#99) Detect FreeBSD and installed GNU Make in build_third_party.sh
Building the third-party dependencies used to fail on FreeBSD,
because the build script just called out to `make` which defaults
to BSD make. Because of quirks, differences and GNU extensions
the build of ffmpeg failed as well as the link stage of giflib.
Refer to #99 for more details in the discussion.

Fix is to detect FreeBSD and check for gnu make being installed.
If that is the case we just set $MAKE to gmake and continue.

This fix does not check for other BSDs such as OpenBSD, DragonflyBSD,
FuryBSD or Darwin, which might also not be using GNU make by default.

Testing and extending is necessary to fix this problem.