Use fetch instead of wget on FreeBSD
Since wget is a GNU tool, it doesn't come by default with
FreeBSD. Instead we wanna use fetch, which will always be available on
that platform. The change just adds another variable that will be set
accordingly whenever a FreeBSD system is discovered. Since it might
also apply to different operating systems, testing may be required.